feat: export octave_filter and octavefilter from phonometry.metrology
Move the octave_filter() convenience wrapper, its deprecated
octavefilter() alias and the design cache from phonometry/__init__.py
into phonometry.metrology.core, next to the OctaveFilterBank they wrap.
Both names are now re-exported by the metrology subpackage (matching
how every other deprecated alias propagates, e.g. normalizedfreq) and
the top level keeps re-exporting them, so existing imports are
unaffected and the deprecation warning behavior is unchanged.

The generated API reference now documents both under metrology.core,
and the API index and llms.txt quickstart present the modular
subpackage import as the primary form.
