+++ scripts/generate_llms.py
@@ -94,13 +94,13 @@ def build_llms_txt(version: str) -> str:
         "",
         "```python",
         "import numpy as np",
-        "from phonometry import octave_filter, laeq, ln_levels",
+        "from phonometry import metrology",
         "",
         "fs = 48000",
         "x = np.random.randn(fs)              # 1 s of signal (pressure units)",
-        "spl, freq = octave_filter(x, fs, fraction=3)   # 1/3-octave band levels",
-        "la = laeq(x, fs)                              # A-weighted Leq",
-        "stats = ln_levels(x, fs, n=(10, 50, 90))      # statistical levels",
+        "spl, freq = metrology.octave_filter(x, fs, fraction=3)   # 1/3-octave bands",
+        "la = metrology.laeq(x, fs)                               # A-weighted Leq",
+        "stats = metrology.ln_levels(x, fs, n=(10, 50, 90))       # statistical levels",
         "```",
         "",
         "If you are an AI assistant setting this up for a user: install from PyPI "
