+++ src/phonometry/metrology/core.py
@@ -6,12 +6,13 @@ Core processing logic and FilterBank class for phonometry.
 from __future__ import annotations
 
 import warnings
-from typing import List, Tuple, cast, overload, Literal
+from functools import lru_cache
+from typing import Any, List, Tuple, cast, overload, Literal
 
 import numpy as np
 from scipy import signal
 
-from .._internal.warnings import PhonometryWarning
+from .._internal.warnings import PhonometryWarning, _warn_renamed
 from .filter_design import _cheby2_headroom, _design_sos_filter
 from .frequencies import _genfreqs
 from .._internal.utils import _downsamplingfactor, _resample_to_length, _typesignal
@@ -489,3 +490,212 @@ class OctaveFilterBank:
         # Physical SPL: apply sensitivity and use 20uPa reference
         pressure_pa = val_linear * self.calibration_factor
         return cast(np.ndarray, 20 * np.log10(np.maximum(pressure_pa, eps) / 2e-5))
+
+
+@lru_cache(maxsize=32)
+def _cached_filter_bank(
+    fs: int,
+    fraction: float,
+    order: int,
+    limits: Tuple[float, ...] | None,
+    filter_type: str,
+    ripple: float,
+    attenuation: float,
+    calibration_factor: float,
+    dbfs: bool,
+) -> OctaveFilterBank:
+    """Design (or reuse) a stateless filter bank for octave_filter()."""
+    return OctaveFilterBank(
+        fs=fs,
+        fraction=fraction,
+        order=order,
+        limits=list(limits) if limits is not None else None,
+        filter_type=filter_type,
+        ripple=ripple,
+        attenuation=attenuation,
+        calibration_factor=calibration_factor,
+        dbfs=dbfs,
+    )
+
+
+@overload
+def octave_filter(
+    x: List[float] | np.ndarray,  # NOSONAR - public API
+    fs: int,
+    fraction: float = 1,
+    order: int = 6,
+    limits: List[float] | None = None,
+    show: bool = False,
+    sigbands: Literal[False] = False,
+    plot_file: str | None = None,
+    detrend: bool = True,
+    filter_type: str = "butter",
+    ripple: float = 0.1,
+    attenuation: float = 72.0,
+    calibration_factor: float = 1.0,
+    dbfs: bool = False,
+    mode: str = "rms",
+    nominal: Literal[False] = False,
+) -> Tuple[np.ndarray, List[float]]: ...
+
+
+@overload
+def octave_filter(
+    x: List[float] | np.ndarray,  # NOSONAR - public API
+    fs: int,
+    fraction: float = 1,
+    order: int = 6,
+    limits: List[float] | None = None,
+    show: bool = False,
+    sigbands: Literal[True] = True,
+    plot_file: str | None = None,
+    detrend: bool = True,
+    filter_type: str = "butter",
+    ripple: float = 0.1,
+    attenuation: float = 72.0,
+    calibration_factor: float = 1.0,
+    dbfs: bool = False,
+    mode: str = "rms",
+    nominal: Literal[False] = False,
+) -> Tuple[np.ndarray, List[float], List[np.ndarray]]: ...
+
+
+@overload
+def octave_filter(
+    x: List[float] | np.ndarray,  # NOSONAR - public API
+    fs: int,
+    fraction: float = 1,
+    order: int = 6,
+    limits: List[float] | None = None,
+    show: bool = False,
+    sigbands: Literal[False] = False,
+    plot_file: str | None = None,
+    detrend: bool = True,
+    filter_type: str = "butter",
+    ripple: float = 0.1,
+    attenuation: float = 72.0,
+    calibration_factor: float = 1.0,
+    dbfs: bool = False,
+    mode: str = "rms",
+    nominal: Literal[True] = ...,
+) -> Tuple[np.ndarray, List[str]]: ...
+
+
+@overload
+def octave_filter(
+    x: List[float] | np.ndarray,  # NOSONAR - public API
+    fs: int,
+    fraction: float = 1,
+    order: int = 6,
+    limits: List[float] | None = None,
+    show: bool = False,
+    sigbands: Literal[True] = True,
+    plot_file: str | None = None,
+    detrend: bool = True,
+    filter_type: str = "butter",
+    ripple: float = 0.1,
+    attenuation: float = 72.0,
+    calibration_factor: float = 1.0,
+    dbfs: bool = False,
+    mode: str = "rms",
+    nominal: Literal[True] = ...,
+) -> Tuple[np.ndarray, List[str], List[np.ndarray]]: ...
+
+
+def octave_filter(
+    x: List[float] | np.ndarray,  # NOSONAR - public API
+    fs: int,
+    fraction: float = 1,
+    order: int = 6,
+    limits: List[float] | None = None,
+    show: bool = False,
+    sigbands: bool = False,
+    plot_file: str | None = None,
+    detrend: bool = True,
+    filter_type: str = "butter",
+    ripple: float = 0.1,
+    attenuation: float = 72.0,
+    calibration_factor: float = 1.0,
+    dbfs: bool = False,
+    mode: str = "rms",
+    nominal: bool = False,
+) -> Tuple[np.ndarray, List[float]] | Tuple[np.ndarray, List[str]] | Tuple[np.ndarray, List[float], List[np.ndarray]] | Tuple[np.ndarray, List[str], List[np.ndarray]]:
+    """
+    Filter a signal with octave or fractional octave filter bank.
+
+    This method uses a filter bank with Second-Order Sections (SOS) coefficients.
+    To obtain the correct coefficients, automatic subsampling is applied to the
+    signal in each filtered band.
+
+    Multichannel support: If x is 2D (channels, samples), each channel is filtered.
+
+    :param x: Input signal (1D array or 2D array [channels, samples]).
+    :type x: Union[List[float], np.ndarray]
+    :param fs: Sample rate in Hz.
+    :type fs: int
+    :param fraction: Bandwidth 'b'. Examples: 1/3-octave b=3, 1-octave b=1, 2/3-octave b=1.5. Default: 1.
+    :type fraction: float
+    :param order: Order of the filter. Default: 6.
+    :type order: int
+    :param limits: Minimum and maximum limit frequencies [f_min, f_max]. Default [12, 20000].
+    :type limits: Optional[List[float]]
+    :param show: If True, plot and show the filter response.
+    :type show: bool
+    :param sigbands: If True, also return the signal in the time domain divided into bands.
+    :type sigbands: bool
+    :param plot_file: Path to save the filter response plot.
+    :type plot_file: Optional[str]
+    :param detrend: If True, remove DC offset before filtering. Default: True.
+    :type detrend: bool
+    :param filter_type: Type of filter ('butter', 'cheby1', 'cheby2', 'ellip', 'bessel').
+        Default: 'butter' (the only type that meets IEC 61260-1 class 1 with the
+        default parameters).
+    :param ripple: Passband ripple in dB (for cheby1, ellip). Default: 0.1.
+    :param attenuation: Stopband attenuation in dB (for cheby2, ellip). Default: 72.0.
+        For ``cheby2`` scipy pins the deep-stopband floor at exactly this value,
+        so it must be >= 70 dB to clear the IEC 61260-1 class 1 limit (matches
+        :class:`OctaveFilterBank`).
+    :param calibration_factor: Calibration factor for SPL calculation. Default: 1.0.
+    :param dbfs: If True, return results in dBFS. Default: False.
+    :param mode: 'rms' or 'peak'. Default: 'rms'.
+    :param nominal: If True, return IEC 61260-1 nominal frequency labels (List[str]) instead of exact floats.
+    :return: A tuple containing (SPL_array, Frequencies_list) or (SPL_array, Frequencies_list, signals).
+        When *nominal=True*, the frequency list contains ``List[str]`` labels instead of floats.
+    :rtype: Union[Tuple[np.ndarray, List[float]], Tuple[np.ndarray, List[str]],
+        Tuple[np.ndarray, List[float], List[np.ndarray]],
+        Tuple[np.ndarray, List[str], List[np.ndarray]]]
+    """
+    
+    if show or plot_file:
+        # Plotting has side effects: bypass the cache.
+        filter_bank = OctaveFilterBank(
+            fs=fs,
+            fraction=fraction,
+            order=order,
+            limits=limits,
+            filter_type=filter_type,
+            ripple=ripple,
+            attenuation=attenuation,
+            show=show,
+            plot_file=plot_file,
+            calibration_factor=calibration_factor,
+            dbfs=dbfs,
+        )
+    else:
+        # The bank is immutable in non-stateful mode: reuse the design.
+        # Pass limits through as-is (tuple for hashability); the bank
+        # constructor is the single place that validates them and owns
+        # the default when None.
+        limits_key = tuple(map(float, limits)) if limits is not None else None
+        filter_bank = _cached_filter_bank(
+            fs, fraction, order, limits_key, filter_type,
+            ripple, attenuation, calibration_factor, dbfs,
+        )
+
+    return filter_bank.filter(x, sigbands=sigbands, mode=mode, detrend=detrend, nominal=nominal)  # type: ignore[call-overload,no-any-return]
+
+
+def octavefilter(*args: Any, **kwargs: Any) -> Any:
+    """Deprecated alias of :func:`octave_filter`."""
+    _warn_renamed("octavefilter()", "octave_filter()")
+    return octave_filter(*args, **kwargs)
